Some improvements

+++ project4-master/objects.c
@@ -12,12 +12,12 @@ void init_game(object court_net, object player, object opponent, object ball){
 	GLCD_Bitmap(court_net.x,court_net.y,court_net.width,court_net.height, court_net.bitmap); // only need to draw court net once 
 	GLCD_Bitmap(opponent.x,opponent.y,opponent.width,opponent.height, opponent.bitmap);
 	GLCD_Bitmap(player.x,player.y,player.width,player.height, player.bitmap);
-    GLCD_Bitmap(ball.x, ball.y, ball.width, ball.height, ball.bitmap);
+  GLCD_Bitmap(ball.x, ball.y, ball.width, ball.height, ball.bitmap);
 }
 
 
 void write_obect_lcd(object obj){
-    GLCD_Bitmap(obj.x,obj.y,obj.width,obj.height, obj.bitmap);
+    GLCD_Bitmap((unsigned int)(obj.x),(unsigned int)(obj.y),obj.width,obj.height, obj.bitmap);
 }
